03-API/trabalho/requirements.txt
fastapi
litellm
uvicorn
PyJWT
PyPDF2
sqlalchemy
passlib
python-multipart